requirements.txt
fastapi
uvicorn[standard]
google-generativeai
chromadb
sentence-transformers
python-dotenv
pydantic